mod-comments: move client-related calls (connect, signal, call, send) to own module
